requirements.txt
# Core Dependencies
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Database
mysql-connector-python==8.1.0

# Authentication
bcrypt==4.0.1

# Machine Learning
scikit-learn>=1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dateutil==2.8.2
